move gen_stat_tables to own module
